backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "learnova-backend"
version = "0.1.0"
description = "Learnova AI LMS backend (FastAPI + Celery)"
authors = [{name = "Learnova"}]
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.6",
    "SQLAlchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "psycopg[binary,pool]>=3.2",
    "qdrant-client>=1.9",
    "celery>=5.3",
    "pika>=1.3",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "tiktoken>=0.7",
    "faster-whisper>=1.0",
    "ffmpeg-python>=0.2",
    "yt-dlp>=2024.4.9",
    "youtube-transcript-api>=0.6",
    "pdfminer.six>=20231228",
    "python-docx>=1.1",
    "python-pptx>=0.6",
    "unstructured[all-docs]>=0.14",
    "pytesseract>=0.3",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "black", "mypy"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pyright]
typeCheckingMode = "basic"
